Implemented minimal connect-only client socket.
Created an `sdk` directory to contain minimalist network programs
written in C++ using the C++ implementation of UDT. I've created a
minimal client and server program to upload a file from the client to
the server. The upload sends only the file, and the server writes the
file to the filesystem using a generated file name. These programs
illustrate client connections and listening. The sender's algorithm is
exercised in the client. The receiver's algorithm is exercised in the
server.

This allows me to build out connecting and sending by implementing the
client in JavaScript, before building out listening and receiving by
implementing the server in JavaScript.

Implemented client connections, sending handshakes repeatedly until
timeout. Implemented shutdown using a shutdown packet.

Updated Packet to Packet 0.0.4 to employ the new and reduced API.

Closes #13.