Send single data packet.
Created a minimal send only client socket that can send a single packet
to a server. Created an integer server to test the client socket. The
integer server accepts a single integer, prints it to standard out, adn
closes the connection.

Gathered the C++ example programs into separate directories so I could
reuse the names `client` and `server` for each pair, and the directory
name to describe the program's operation.

Implemented a binary heap. Has strange premonition of admonishments over
the use of socket object properties in the heap implementations. It may
offend the sensibilities of certain delicate souls because it uses a
property in the object that it stores to store a bit of its
implementation data &mdash; the array index.

It creates an arbitrary property name to store the index and deletes the
property when the object is removed. It knows little about the stored
object, just that it is an object and not a primitive, and it has some
sense of what sort of property name will not collide with the properties
already defined in the object.

The name binary heap uses a comparator function, as you might expect of
a generalized data structure. The difference being that the heap uses a
stored object property to store its data, establishing the requirement
that the stored type must be of type object, but it does not know
anything more about the object. It defers to a comparator to define the
ordering.

There is a lot more to do to implement the senders algorithm, but that
would be a huge commits that would bury the incremental progress being
made. Now that I have a better understanding of how the protocol works,
I can better define the steps to implement it.

Closes #9.