Implement light acknowledgement.
Implement the acknowledgement acknowledgement in response to an
acknowledgement. This works for a single packet. The `client.js` program
correctly connects to the C++ implementation of the integer server,
sends and integer, and acknowledges the server's acknowledgement of
packet receipt.

Added an optional five second sleep to the integer server's response
thread between the socket read and the socket close to give the socket a
chance to preform the acknowledgement exchange before hanging up.
Without it, when the C++ implementation of the client is used, a
shutdown is sent to the client before it has a chance to send an
acknowledgement. The option is tiggered by a command line argument.

Closes #26.