Wrap sequence number.
Wrap the socket's packet sequence number when it reaches the maximum
sequence number value. Track the sequences and maintain the order in
which they were sent across the wrap boundary. That is, when we receive
an acknowledgement packet, we know that acknowledging packet 2 would
mean that packet 2147483647 has also been acknowledged.

Closes #24.