Create `Makefile` to build UDT4 and C++ examples.
Created a `Makefile` that will download and build the UDT4 C++ reference
implementation and build it along with our C++ example programs. We're
going to use the server side of the C++ programs as part of our strategy
to ensure that we conform to the protocol, testing against the UDT4 C++
reference on the platforms where it will easily compile, ensuring
protocol parity between the C++ and JavaScript implementations.